Count recomputed 0/0 genotypes with alt reads > 0

+++ vcf_sample_counter_merger.sh
@@ -3,7 +3,7 @@
 
 
 if [ $# -eq 0 ]; then
-	echo "usage : this_script.sh db.vcf  otherVCF1.vcf otherVCF2.vcf otherVCF3.vcf otherVCF4.vcf otherVCF5.vcf ...... " 
+	echo "usage : this_script.sh outputFileName.vcf  VCF1.vcf VCF2.vcf VCF3.vcf VCF4.vcf VCF5.vcf ...... " 
     echo "No arguments provided"
     exit 1
 fi
@@ -14,20 +14,21 @@ echo "it will output in "$1
 
 
 
-echo "Copy first argument as backup "  
+echo "Backup outputFilename if file exists"  
 #datestamp=`date +%F`
 printf -v date '%(%Y-%m-%d_at_%Hh%Mm%Ss)T\n' -1
-cp $1 ${1}.$date
+if [ -f "$1" ]; then
+	cp $1 ${1}.$date
+fi
 
 
-
-awk -F "\t" 'BEGIN{SAMPLES="";found=0;samplesFound="";HMZ=0;HTZ=0;TOTALSAMPLES=0;CHROMHEADER="";OFS="\t"}
+awk -F "\t" 'BEGIN{SAMPLES="";found=0;samplesFound="";HMZ=0;HTZ=0;RECOMPUTED=0;TOTALSAMPLES=0;CHROMHEADER="";OFS="\t"}
 {	if ($0 ~ /^##/){
 		if(SAMPLES==""){print};next;
 	}
 
 	if ($0 ~ /^#CHROM/){
-				CHROMHEADER= $1"\t"$2"\t"$3"\t"$4"\t"$5"\t"$6"\t"$7"\t"$8;	
+		CHROMHEADER= $1"\t"$2"\t"$3"\t"$4"\t"$5"\t"$6"\t"$7"\t"$8;	
 
 		split($0,sample,"\t")
 		for (i=10;i<=NF; i++){
@@ -42,14 +43,16 @@ awk -F "\t" 'BEGIN{SAMPLES="";found=0;samplesFound="";HMZ=0;HTZ=0;TOTALSAMPLES=0
 	}else{ 
 			VARTAB[$1"_"$2"_"$4"_"$5]["init"]=$0;
 			VARTAB[$1"_"$2"_"$4"_"$5]["startHalf"]=$1"\t"$2"\t"$3"\t"$4"\t"$5"\t"$6"\t"$7;
-			
+
 			if($8~/^found=/){
 				gsub("=|;","\t",$8);
 				split($8,INFO,"\t");
 				found=INFO[2];
 				HTZ=INFO[4];
 				HMZ=INFO[6];
-				samplesFound=INFO[8];
+				RECOMPUTED=INFO[8];
+				samplesFound=INFO[10];
+				SAMPLES="DBfound";
 	
 			}else{
 
@@ -66,7 +69,16 @@ awk -F "\t" 'BEGIN{SAMPLES="";found=0;samplesFound="";HMZ=0;HTZ=0;TOTALSAMPLES=0
 							found++;
 							HMZ++;
 							samplesFound=sample[i]"/"samplesFound;
+						}else if($i ~ /^0\/0/ ){
+							split($i,recomputed,":")
+							split(recomputed[2],altdepth,",")
+							if (altdepth[2]>=1){
+								found++;
+								RECOMPUTED++;
+								samplesFound=sample[i]"/"samplesFound;
+							}
 						}
+
 					}
 				}
 			}
@@ -75,14 +87,16 @@ awk -F "\t" 'BEGIN{SAMPLES="";found=0;samplesFound="";HMZ=0;HTZ=0;TOTALSAMPLES=0
 			VARTAB[$1"_"$2"_"$4"_"$5]["found"] += found;
 			VARTAB[$1"_"$2"_"$4"_"$5]["HTZ"] += HTZ;
 			VARTAB[$1"_"$2"_"$4"_"$5]["HMZ"] += HMZ;
+			VARTAB[$1"_"$2"_"$4"_"$5]["RECOMPUTED"] += RECOMPUTED;
 			VARTAB[$1"_"$2"_"$4"_"$5]["samplesFound"] = VARTAB[$1"_"$2"_"$4"_"$5]["samplesFound"]samplesFound;
   
-			VARTAB[$1"_"$2"_"$4"_"$5]["INFO"]="found="VARTAB[$1"_"$2"_"$4"_"$5]["found"]";HTZ="VARTAB[$1"_"$2"_"$4"_"$5]["HTZ"]";HMZ="VARTAB[$1"_"$2"_"$4"_"$5]["HMZ"]";samplesFound="VARTAB[$1"_"$2"_"$4"_"$5]["samplesFound"];
+			VARTAB[$1"_"$2"_"$4"_"$5]["INFO"]="found="VARTAB[$1"_"$2"_"$4"_"$5]["found"]";HTZ="VARTAB[$1"_"$2"_"$4"_"$5]["HTZ"]";HMZ="VARTAB[$1"_"$2"_"$4"_"$5]["HMZ"]";RECOMPUTED="VARTAB[$1"_"$2"_"$4"_"$5]["RECOMPUTED"]";samplesFound="VARTAB[$1"_"$2"_"$4"_"$5]["samplesFound"];
 		
 			found=0;
 			samplesFound="";
 			HMZ=0;
 			HTZ=0;
+			RECOMPUTED=0;
 		}
 }
